ModelManager: moved all methods into Model as it can manage itself
